+++ stacks-signer/src/v0/signer.rs
@@ -45,13 +45,15 @@ use stacks_common::util::secp256k1::MessageSignature;
 use stacks_common::{debug, error, info, warn};
 
 use super::signer_state::LocalStateMachine;
-#[cfg(not(any(test, feature = "testing")))]
-use super::signer_state::SUPPORTED_SIGNER_PROTOCOL_VERSION;
-use crate::chainstate::{ProposalEvalConfig, SortitionMinerStatus, SortitionsView};
+use crate::chainstate::v1::{SortitionMinerStatus, SortitionsView};
+use crate::chainstate::v2::GlobalStateView;
+use crate::chainstate::{ProposalEvalConfig, SortitionData, SortitionStateVersion};
 use crate::client::{ClientError, SignerSlotID, StackerDB, StacksClient};
 use crate::config::{SignerConfig, SignerConfigMode};
 use crate::runloop::SignerResult;
 use crate::signerdb::{BlockInfo, BlockState, SignerDb};
+#[cfg(not(any(test, feature = "testing")))]
+use crate::v0::signer_state::SUPPORTED_SIGNER_PROTOCOL_VERSION;
 use crate::v0::signer_state::{NewBurnBlock, ReplayScopeOpt};
 use crate::Signer as SignerTrait;
 
@@ -127,6 +129,11 @@ pub struct Signer {
     pub validate_with_replay_tx: bool,
     /// Scope of Tx Replay in terms of Burn block boundaries
     pub tx_replay_scope: ReplayScopeOpt,
+    /// Time to wait between updating our local state machine view point and capitulating to other signers miner view
+    pub capitulate_miner_view_timeout: Duration,
+    /// The signer supported protocol version. used only in testing
+    #[cfg(any(test, feature = "testing"))]
+    pub supported_signer_protocol_version: u64,
 }
 
 impl std::fmt::Display for SignerMode {
@@ -204,11 +211,6 @@ impl SignerTrait<SignerMessage> for Signer {
             SignerDb::new(&signer_config.db_path).expect("Failed to connect to signer Db");
         let proposal_config = ProposalEvalConfig::from(&signer_config);
 
-        let signer_state = LocalStateMachine::new(&signer_db, stacks_client, &proposal_config)
-            .unwrap_or_else(|e| {
-                warn!("Failed to initialize local state machine for signer: {e:?}");
-                LocalStateMachine::Uninitialized
-            });
         let stacks_address = StacksAddress::p2pkh(
             signer_config.mainnet,
             &StacksPublicKey::from_private(&signer_config.stacks_private_key),
@@ -224,6 +226,21 @@ impl SignerTrait<SignerMessage> for Signer {
             updates,
             signer_config.signer_entries.signer_addr_to_weight.clone(),
         );
+        #[cfg(any(test, feature = "testing"))]
+        let version = signer_config.supported_signer_protocol_version;
+        #[cfg(not(any(test, feature = "testing")))]
+        let version = SUPPORTED_SIGNER_PROTOCOL_VERSION;
+        let signer_state = LocalStateMachine::new(
+            &signer_db,
+            stacks_client,
+            &proposal_config,
+            &global_state_evaluator,
+            version,
+        )
+        .unwrap_or_else(|e| {
+            warn!("Failed to initialize local state machine for signer: {e:?}");
+            LocalStateMachine::Uninitialized
+        });
         Self {
             private_key: signer_config.stacks_private_key,
             stacks_address,
@@ -244,6 +261,9 @@ impl SignerTrait<SignerMessage> for Signer {
             global_state_evaluator,
             validate_with_replay_tx: signer_config.validate_with_replay_tx,
             tx_replay_scope: None,
+            capitulate_miner_view_timeout: signer_config.capitulate_miner_view_timeout,
+            #[cfg(any(test, feature = "testing"))]
+            supported_signer_protocol_version: signer_config.supported_signer_protocol_version,
         }
     }
 
@@ -265,12 +285,24 @@ impl SignerTrait<SignerMessage> for Signer {
         self.check_pending_block_validations(stacks_client);
 
         let mut prior_state = self.local_state_machine.clone();
+        let local_signer_protocol_version = self.get_signer_protocol_version();
         if self.reward_cycle <= current_reward_cycle {
             self.local_state_machine.handle_pending_update(&self.signer_db, stacks_client,
                 &self.proposal_config,
-                &mut self.tx_replay_scope)
+                &mut self.tx_replay_scope, &self.global_state_evaluator, local_signer_protocol_version)
                 .unwrap_or_else(|e| error!("{self}: failed to update local state machine for pending update"; "err" => ?e));
         }
+        // See if we should capitulate our viewpoint...
+        self.local_state_machine.capitulate_viewpoint(
+            stacks_client,
+            &mut self.signer_db,
+            &mut self.global_state_evaluator,
+            local_signer_protocol_version,
+            self.reward_cycle,
+            sortition_state,
+            self.capitulate_miner_view_timeout,
+            self.proposal_config.tenure_last_block_proposal_timeout,
+        );
 
         if prior_state != self.local_state_machine {
             let version = self.get_signer_protocol_version();
@@ -435,12 +467,7 @@ impl Signer {
                             sortition_state,
                         ),
                         SignerMessage::StateMachineUpdate(update) => self
-                            .handle_state_machine_update(
-                                signer_public_key,
-                                update,
-                                received_time,
-                                sortition_state,
-                            ),
+                            .handle_state_machine_update(signer_public_key, update, received_time),
                         SignerMessage::BlockPreCommit(block_pre_commit) => {
                             if let Ok(pubkey) = block_pre_commit.recover_public_key(self.mainnet).inspect_err(|e| warn!("Unable to recover public key from block pre commit {block_pre_commit:?}: {e}")) {
                                 let stacker_address = StacksAddress::p2pkh(self.mainnet, &pubkey);
@@ -467,19 +494,10 @@ impl Signer {
                             if self.test_ignore_all_block_proposals(block_proposal) {
                                 continue;
                             }
-                            let Some(miner_pubkey) = block_proposal.block.header.recover_miner_pk()
-                            else {
-                                warn!("{self}: Failed to recover miner pubkey";
-                                      "signer_signature_hash" => %block_proposal.block.header.signer_signature_hash(),
-                                      "consensus_hash" => %block_proposal.block.header.consensus_hash);
-                                continue;
-                            };
-
                             self.handle_block_proposal(
                                 stacks_client,
                                 sortition_state,
                                 block_proposal,
-                                &miner_pubkey,
                             );
                         }
                         SignerMessage::BlockPushed(b) => {
@@ -548,13 +566,14 @@ impl Signer {
                         panic!("{self} Failed to write burn block event to signerdb: {e}");
                     });
 
+                let active_signer_protocol_version = self.get_signer_protocol_version();
                 self.local_state_machine
                     .bitcoin_block_arrival(&self.signer_db, stacks_client, &self.proposal_config, Some(NewBurnBlock {
                         burn_block_height: *burn_height,
                         consensus_hash: *consensus_hash,
                     }),
                     &mut self.tx_replay_scope
-                )
+                , &self.global_state_evaluator, active_signer_protocol_version)
                     .unwrap_or_else(|e| error!("{self}: failed to update local state machine for latest bitcoin block arrival"; "err" => ?e));
                 *sortition_state = None;
             }
@@ -647,14 +666,41 @@ impl Signer {
         peer_info.stacks_tip_height >= block.header.chain_length.saturating_sub(1)
     }
 
-    /// Check if block should be rejected based on sortition state
+    /// Check if block should be rejected based on the appropriate state (either local or global)
     /// Will return a BlockResponse::Rejection if the block is invalid, none otherwise.
-    fn check_block_against_sortition_state(
+    fn check_block_against_state(
         &mut self,
         stacks_client: &StacksClient,
         sortition_state: &mut Option<SortitionsView>,
         block: &NakamotoBlock,
-        miner_pubkey: &Secp256k1PublicKey,
+    ) -> Option<BlockResponse> {
+        let Some(latest_version) = self
+            .global_state_evaluator
+            .determine_latest_supported_signer_protocol_version()
+        else {
+            warn!(
+                "{self}: No consensus on signer protocol version. Unable to validate block. Rejecting.";
+                "signer_signature_hash" => %block.header.signer_signature_hash(),
+                "block_id" => %block.block_id(),
+            );
+            return Some(self.create_block_rejection(RejectReason::NoSignerConsensus, block));
+        };
+        let state_version = SortitionStateVersion::from_protocol_version(latest_version);
+        if state_version.uses_global_state() {
+            self.check_block_against_global_state(stacks_client, block)
+        } else {
+            self.check_block_against_local_state(stacks_client, sortition_state, block)
+        }
+    }
+
+    /// Check if block should be rejected based on the local view of the sortition state
+    /// Will return a BlockResponse::Rejection if the block is invalid, none otherwise.
+    /// This is the pre-global signer state activation path.
+    fn check_block_against_local_state(
+        &mut self,
+        stacks_client: &StacksClient,
+        sortition_state: &mut Option<SortitionsView>,
+        block: &NakamotoBlock,
     ) -> Option<BlockResponse> {
         let signer_signature_hash = block.header.signer_signature_hash();
         let block_id = block.block_id();
@@ -674,13 +720,7 @@ impl Signer {
 
         // Check if proposal can be rejected now if not valid against sortition view
         if let Some(sortition_state) = sortition_state {
-            match sortition_state.check_proposal(
-                stacks_client,
-                &mut self.signer_db,
-                block,
-                miner_pubkey,
-                true,
-            ) {
+            match sortition_state.check_proposal(stacks_client, &mut self.signer_db, block, true) {
                 // Error validating block
                 Err(RejectReason::ConnectivityIssues(e)) => {
                     warn!(
@@ -711,6 +751,76 @@ impl Signer {
                 "block_id" => %block_id,
             );
             Some(self.create_block_rejection(RejectReason::NoSortitionView, block))
+        }
+    }
+
+    /// Check if block should be rejected based on global signer state
+    /// Will return a BlockResponse::Rejection if the block is invalid, none otherwise.
+    /// This is the Post-global signer state activation path
+    fn check_block_against_global_state(
+        &mut self,
+        stacks_client: &StacksClient,
+        block: &NakamotoBlock,
+    ) -> Option<BlockResponse> {
+        let signer_signature_hash = block.header.signer_signature_hash();
+        let block_id = block.block_id();
+        // First update our global state evaluator with our local state if we have one
+        let version = self.get_signer_protocol_version();
+        if let Ok(update) = self
+            .local_state_machine
+            .try_into_update_message_with_version(version)
+        {
+            self.global_state_evaluator
+                .insert_update(self.stacks_address, update);
+        };
+
+        let Some(global_state) = self.global_state_evaluator.determine_global_state() else {
+            warn!(
+                "{self}: Cannot validate block, no global signer state";
+                "signer_signature_hash" => %signer_signature_hash,
+                "block_id" => %block_id,
+                "local_signer_state" => ?self.local_state_machine
+            );
+            return Some(self.create_block_rejection(RejectReason::NoSignerConsensus, block));
+        };
+
+        let global_state_view = GlobalStateView {
+            signer_state: global_state,
+            config: self.proposal_config.clone(),
+        };
+
+        info!(
+            "{self}: Evaluating proposal against global state";
+            "signer_state" => ?global_state_view.signer_state,
+            "signer_signature_hash" => %signer_signature_hash,
+            "block_id" => %block_id,
+            "local_signer_state" => ?self.local_state_machine,
+        );
+
+        // Check if proposal can be rejected now if not valid against the global state
+        match global_state_view.check_proposal(stacks_client, &mut self.signer_db, block) {
+            // Error validating block
+            Err(RejectReason::ConnectivityIssues(e)) => {
+                warn!(
+                    "{self}: Error checking block proposal: {e}";
+                    "signer_signature_hash" => %signer_signature_hash,
+                    "block_id" => %block_id,
+                );
+                Some(self.create_block_rejection(RejectReason::ConnectivityIssues(e), block))
+            }
+            // Block proposal is bad
+            Err(reject_code) => {
+                warn!(
+                    "{self}: Block proposal invalid";
+                    "signer_signature_hash" => %signer_signature_hash,
+                    "block_id" => %block_id,
+                    "reject_reason" => %reject_code,
+                    "reject_code" => ?reject_code,
+                );
+                Some(self.create_block_rejection(reject_code, block))
+            }
+            // Block proposal passed check, still don't know if valid
+            Ok(_) => None,
         }
     }
 
@@ -798,8 +908,8 @@ impl Signer {
         signer_public_key: &Secp256k1PublicKey,
         update: &StateMachineUpdate,
         received_time: &SystemTime,
-        sortition_state: &mut Option<SortitionsView>,
     ) {
+        info!("{self}: Received a new state machine update from signer {signer_public_key:?}: {update:?}");
         let address = StacksAddress::p2pkh(self.mainnet, signer_public_key);
         // Store the state machine update so we can reload it if we crash
         if let Err(e) = self.signer_db.insert_state_machine_update(
@@ -812,17 +922,6 @@ impl Signer {
         }
         self.global_state_evaluator
             .insert_update(address, update.clone());
-
-        // See if this update means we should capitulate our viewpoint...
-        let version = self.get_signer_protocol_version();
-        self.local_state_machine.capitulate_viewpoint(
-            &mut self.signer_db,
-            &mut self.global_state_evaluator,
-            self.stacks_address,
-            version,
-            self.reward_cycle,
-            sortition_state,
-        );
     }
 
     /// Handle pre-commit message from another signer
@@ -920,7 +1019,6 @@ impl Signer {
         stacks_client: &StacksClient,
         sortition_state: &mut Option<SortitionsView>,
         block_proposal: &BlockProposal,
-        miner_pubkey: &Secp256k1PublicKey,
     ) {
         debug!("{self}: Received a block proposal: {block_proposal:?}");
         if block_proposal.reward_cycle != self.reward_cycle {
@@ -997,12 +1095,8 @@ impl Signer {
         }
 
         // Check if proposal can be rejected now if not valid against sortition view
-        let block_response = self.check_block_against_sortition_state(
-            stacks_client,
-            sortition_state,
-            &block_proposal.block,
-            miner_pubkey,
-        );
+        let block_response =
+            self.check_block_against_state(stacks_client, sortition_state, &block_proposal.block);
 
         #[cfg(any(test, feature = "testing"))]
         let block_response =
@@ -1097,7 +1191,7 @@ impl Signer {
         // If this is a tenure change block, ensure that it confirms the correct number of blocks from the parent tenure.
         if let Some(tenure_change) = proposed_block.get_tenure_change_tx_payload() {
             // Ensure that the tenure change block confirms the expected parent block
-            match SortitionsView::check_tenure_change_confirms_parent(
+            match SortitionData::check_tenure_change_confirms_parent(
                 tenure_change,
                 proposed_block,
                 &mut self.signer_db,
@@ -1128,7 +1222,7 @@ impl Signer {
         }
 
         // Ensure that the block is the last block in the chain of its current tenure.
-        match SortitionsView::check_latest_block_in_tenure(
+        match SortitionData::check_latest_block_in_tenure(
             &proposed_block.header.consensus_hash,
             proposed_block,
             &mut self.signer_db,
@@ -1551,6 +1645,7 @@ impl Signer {
             self.submitted_block_proposal = None;
         }
 
+        // NOTE: This is only used by active signer protocol versions < 2
         // If 30% of the signers have rejected the block due to an invalid
         // reorg, mark the miner as invalid.
         let total_reorg_reject_weight = self.compute_reject_code_signing_weight(
@@ -1561,7 +1656,7 @@ impl Signer {
             // Mark the miner as invalid
             if let Some(sortition_state) = sortition_state {
                 let ch = block_info.block.header.consensus_hash;
-                if sortition_state.cur_sortition.consensus_hash == ch {
+                if sortition_state.cur_sortition.data.consensus_hash == ch {
                     info!("{self}: Marking miner as invalid for attempted reorg");
                     sortition_state.cur_sortition.miner_status =
                         SortitionMinerStatus::InvalidatedBeforeFirstBlock;
@@ -1833,12 +1928,20 @@ impl Signer {
 
     #[cfg(not(any(test, feature = "testing")))]
     fn get_signer_protocol_version(&self) -> u64 {
-        SUPPORTED_SIGNER_PROTOCOL_VERSION
+        crate::v0::signer_state::SUPPORTED_SIGNER_PROTOCOL_VERSION
     }
 
     #[cfg(any(test, feature = "testing"))]
     fn get_signer_protocol_version(&self) -> u64 {
-        self.test_get_signer_protocol_version()
+        use crate::v0::tests::TEST_PIN_SUPPORTED_SIGNER_PROTOCOL_VERSION;
+        let public_keys = TEST_PIN_SUPPORTED_SIGNER_PROTOCOL_VERSION.get();
+        if let Some(version) = public_keys.get(
+            &stacks_common::types::chainstate::StacksPublicKey::from_private(&self.private_key),
+        ) {
+            warn!("{self}: signer version is pinned to {version}");
+            return *version;
+        }
+        self.supported_signer_protocol_version
     }
 }
 
@@ -1851,6 +1954,8 @@ fn should_reevaluate_block(block_info: &BlockInfo) -> bool {
             | RejectReason::ConnectivityIssues(_)
             | RejectReason::TestingDirective
             | RejectReason::InvalidTenureExtend
+            | RejectReason::ConsensusHashMismatch { .. }
+            | RejectReason::NoSignerConsensus
             | RejectReason::NotRejected
             | RejectReason::Unknown(_) => true,
             RejectReason::ValidationFailed(_)
@@ -1862,7 +1967,8 @@ fn should_reevaluate_block(block_info: &BlockInfo) -> bool {
             | RejectReason::InvalidMiner
             | RejectReason::NotLatestSortitionWinner
             | RejectReason::InvalidParentBlock
-            | RejectReason::DuplicateBlockFound => {
+            | RejectReason::DuplicateBlockFound
+            | RejectReason::IrrecoverablePubkeyHash => {
                 // No need to re-validate these types of rejections.
                 false
             }